@@ -11,23 +11,25 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular/forms": "^6.0.0",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "@angular/router": "^6.0.0",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
+    "@angular/platform-browser": "^6.1.10",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/router": "^6.1.10",
     "core-js": "^2.5.5",
-    "rxjs": "^6.1.0",
+    "rxjs": "^6.3.3",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@types/node": "^10.0.4",
+    "@types/core-js": "^2.5.0",
+    "@types/node": "^10.11.7",
     "angular2-template-loader": "^0.6.2",
     "html-webpack-plugin": "^3.2.0",
     "raw-loader": "^0.5.1",
-    "ts-loader": "^4.3.0",
+    "ts-loader": "^4.5.0",
     "typescript": "^2.8.3",
     "webpack": "4.8.1",
     "webpack-cli": "^2.1.3",
