@@ -1,1 +1,1 @@
-﻿export * from './user';
+﻿export * from './user.interface';
