@@ -1,8 +1,9 @@
-﻿export class User {
+﻿export interface Users {
     id: number;
     username: string;
     password: string;
     famille: string;
+    nourriture: string;
     age: number;
     couleur: string;
     friendsWithMe: string;
