@@ -5,4 +5,6 @@
     templateUrl: 'app.component.html'
 })
 
-export class AppComponent { }
+export class AppComponent {
+
+}
