@@ -2,29 +2,44 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { ReactiveFormsModule }    from '@angular/forms';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-
-// used to create fake backend
-
+import { HttpModule } from '@angular/http';
+import { FormsModule } from '@angular/forms';
 import { AppComponent }  from './app.component';
 import { routing }        from './app.routing';
 
+import { AlertComponent } from './_directives';
+import { AuthGuard } from './_guards';
+import { AlertService, AuthenticationService, UserService } from './_services';
 import { JwtInterceptor, ErrorInterceptor } from './_helpers';
 import { HomeComponent } from './home';
 import { LoginComponent } from './login';
+import { RegisterComponent } from './register';
+import { UsersComponent } from './user/users.component';
+import {FilterPipe} from "./filter.pipe";
 
 @NgModule({
     imports: [
         BrowserModule,
         ReactiveFormsModule,
         HttpClientModule,
-        routing
+        routing,
+        HttpModule,
+        FormsModule,
     ],
     declarations: [
         AppComponent,
+        AlertComponent,
         HomeComponent,
-        LoginComponent
+        LoginComponent,
+        UsersComponent,
+        FilterPipe,
+        RegisterComponent
     ],
     providers: [
+        UserService,
+        AuthGuard,
+        AlertService,
+        AuthenticationService,
         { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
         { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
     ],
